fix(ci): pay the ratchets this branch moved, in the way both prescribe
Two blocking ratchets and neither has a number to edit.

Source size: turn_end_gate_hook.ts crossed the 1500-line ceiling, where
every line is one unit of debt, so the branch came in at +15. Detector E
moved out into _lib/dropped_decision.ts — 139 lines and under the cap, so
its 114-line doc block and function now cost nothing. It moved on merit
rather than only on arithmetic: E is the one detector reading a LIST of
the turn's assistant texts rather than the single closing one, it shares
no input with A-D and F, and its dependencies are the reply-consistency
parser rather than anything the hook owns. The hook re-exports it, so
every existing importer is unchanged. 1515 -> 1400, total back to 17947.

Preamble payload: Iron Law 3 is paid per spawn and arrived at +404 tok
against zero headroom. Its explanation, the measured failure and the
carrier note migrated to the mechanics file the rule already routes to,
which is what the gate's own message prescribes; the obligation stays in
the rule. Three pre-existing prose blocks were tightened to cover the
rest — the 30-session audit paragraph, question pacing, and the opening.
Net delivered-token change is now zero.
