+++ src/domains/meta/pack.yaml
@@ -101,6 +101,6 @@ token_passport:
   catalog_tokens: 3649
   commands_tokens: 236051
   other_tokens: 1298
-  rules_tokens: 65941
-  total_tokens: 306939
+  rules_tokens: 65546
+  total_tokens: 306544
 trust_level_default: core
